Add support for memstick images
Rework the global variables that drive image creation
to accommodate building memstick images. Also allows for
memstick images as input.

New Argument Added:
 -U
	Build memstick images. Check out the updated
	manpage for more information.

New functionality is mostly visible in the $makescript variable,
along with all of its friends. At this time iso and img types
should not be mixed. (so don't use packmule to convert one format
to another)

This should hopefully be wrapped into a new release if anyone is
interested in pushing the FreeBSD port into the tree.