Install packages onto the live cd image
It is useful to add software to a live CD for testing,
like adding DRM graphics drivers to a USB stick to test
on a laptop. This can be done with the LIVE_CD_* configuration
variables. For more information see the man page examples

This is performed by duplicating logic, but performing on
the $workroot directory instead of $pkgroot